small fixes

+++ Geometrijski_Algoritmi/pomocnefunkcije.cpp
@@ -1,12 +1,12 @@
 #include "pomocnefunkcije.h"
 
-double pomocneFunkcije::povrsinaTrougla(const QPoint& A, const QPoint& B, const QPoint& C)
+int pomocneFunkcije::povrsinaTrougla(const QPoint& A, const QPoint& B, const QPoint& C)
 {
     // returns P/2
     return (B.x() - A.x())*(C.y() - A.y()) - (C.x() - A.x())*(B.y() - A.y());
 }
 
-double pomocneFunkcije::distanceKvadrat(const QPoint& A, const QPoint& B)
+int pomocneFunkcije::distanceKvadrat(const QPoint& A, const QPoint& B)
 {
     return (A.x() - B.x())*(A.x() - B.x()) + (A.y() - B.y())*(A.y() - B.y());
 }
