+++ Geometrijski_Algoritmi/pomocnefunkcije.h
@@ -4,8 +4,8 @@
 #include<QPoint>
 
 namespace pomocneFunkcije {
-double povrsinaTrougla(const QPoint& A, const QPoint& B, const QPoint& C);
-double distanceKvadrat(const QPoint& A, const QPoint& B);
+int povrsinaTrougla(const QPoint& A, const QPoint& B, const QPoint& C);
+int distanceKvadrat(const QPoint& A, const QPoint& B);
 }
 
 #endif // POMOCNEFUNKCIJE_H
